1. memindahkan search movie ke file lain 2. membuat navbar 3. membuat route aplikasi 4. membuat landing page sederhana
