+++ package.json
@@ -17,6 +17,7 @@
     "react": "^18.2.0",
     "react-bootstrap": "^2.10.1",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.2",
     "swiper": "^11.0.7"
   },
   "devDependencies": {
