+++ src/App.css
@@ -1,0 +1,7 @@
+.hero-bg {
+  width: 100%;
+  height: auto;
+  background-image: url("./assets/img-hero.jpeg");
+  background-position: top;
+  box-shadow: 0px 0px 10px 20px black;
+}
