--- src/App.jsx
@@ -1,16 +1,0 @@
-import "bootstrap/dist/css/bootstrap.min.css";
-import "bootstrap/dist/js/bootstrap.min.js";
-
-import Homes from "./pages/homes.jsx";
-
-function App() {
-  return (
-    <>
-      <div>
-        <Homes />
-      </div>
-    </>
-  );
-}
-
-export default App;
