--- src/index.css
@@ -1,3 +1,0 @@
-body {
-  background-color: #231c1c;
-}
